LaTeX equation not rendering (#62)
Hey! I was reading this to prepare for the HoTT summer school and noticed an equation wasn't rendering properly. Kind of a bummer since it was the very last equation which the whole page was building up to lol. I figured it's probably because of the line break in between the dollar signs, so I went and removed it. I'm not sure it works because I can't test it right now but thought I should let you know.