Opfibrations (#179)
# Description

This PR is a bit beefy, so it's best to think of it in chunks. First, it
proves a bunch of duality properties of morphism classes (IE: monos in
`C^op` are epis in `C`, etc). Next, it defines the total opposite
category, and proves duality properties for *displayed* morphism
classes. Finally, it defines cocartesian maps and opfibrations, and
proves a bunch of properties about them, mostly using duality arguments.