Bifibrations (#184)
# Description

This PR defines bifibrations, and characterizes them as cartesian
fibrations with left adjoints to base change.
Doing so cleanly requires a minor refactor of some properties of weak
cartesian morphisms, along with some new
definitions:
* Displayed hom categories
* Opreindexing for opfibrations

This PR also adds the (start) of a natural isomorphism API; we should
improve this in the future!

## Notes
I couldn't come up with a better name for `push-out`. Suggestions would
be appreciated!