Use rem for managing layout
While ch and em are conceptually much nicer, the fact that they rely on
the current element's font size makes them less useful in managing
layout as the units are not the same across the page. This causes a
couple of problems:

 - @media breakpoints are entirely incorrect.
 - The sidebar and main contents have different padding, which makes
   them look slightly wonky.

Instead we use rem universally. This does slightly adjust the layout of
the page, but it should now be better behaved.