Build shakefile with default of `-A128M` (#220)
This updates the GHC flags to match those uses in the cabal file, namely
adding `-ATM128` to the default RTS options. This significantly reduces
the time spent in GC.

We could probably set this a little lower if preferred. There are
diminishing returns, so using `-A16M` gives ~44% speedup (compared with
the default `-A4M`), while `-A128M` is maybe 53% faster - the numbers
are a little fuzzy here.