def: macros for creating copattern definitions (#376)
There are a couple of places in the 1Lab where we manually eta-expand
out copattern definitions by hand to get better goal display. This
mostly comes up when dealing with things like subcategories/forgetful
functors; Agda will very happily unfold your nicely named category into
`Restrict Blah`, which is not particularly helpful!

Manually performing these copattern expansions is a bit of a pain, so
this PR adds a small macro that performs this mechanical busywork for
us. This removes the need for the `declare-concrete-category` macro in
#375.