Refactor `Const` (#408)
As noted by @ncfavier, there are too many ways to write cones! This PR
attempts to remove some of this duplication by
redefining `Const` as `!Const . !F`, and banishing `counit->cone`. I've
also done a bit of a naming pass, and renamed `const!` to `!Const`
(among other things) so it doesn't look like a macro invocation.