Free categories, categories from generators and relations (#422)
This PR proves that free categories are, in fact, free. It also proves
that we can construct categories via generators and relations in the
style of Borceux 5.1.6, and shows that these categories are also a free
construction. I've also factored out the definition of graphs into its
own file.

This closes 5.1.2, 5.1.5, and 5.1.6 of Borceux.