Beck-Chevalley conditions (#466)
# Description

This PR defines a general notion of a Beck-Chevalley condition over a
commutative square, and proves that it
is equivalent to the usual definition that involves the invertibility of
a comparison map. Unfortunately, suffering
is conserved, so all of the usual annoying proofs about comparison maps
got compressed into a pair of _very_
annoying proofs, but at least this only needs to be done once!

This is intended to be used in #456, and should be merged first.

## Checklist

Before submitting a merge request, please check the items below:

- [x] I've read [the contributing
guidelines](https://github.com/plt-amy/1lab/blob/main/CONTRIBUTING.md).
- [x] The imports of new modules have been sorted with
`support/sort-imports.hs` (or `nix run --experimental-features
nix-command -f . sort-imports`).
- [x] All new code blocks have "agda" as their language.

If your change affects many files without adding substantial content,
and
you don't want your name to appear on those pages (for example, treewide
refactorings or reformattings), start the commit message and PR title
with `chore:`.

---------

Co-authored-by: Amélia Liao <me@amelia.how>