Fix typoes in Cat.Monoidal.Instances.Day
There are some confusions in the coend diagram on this page trying to
match notation both from the previous equation (of a convolution) and
the following agda equations (between F,G/X,Y and x/c).

I've chosen to amend the diagram to completely match the one before,
since the prose seems to be primarily attempting to compare the two.