Right Kan extensions in Eilenberg-Moore categories
This defines lifting and creation of Kan extensions and shows that right
Kan extensions in Eilenberg-Moore categories are created by the
forgetful functor, generalising the creation of limits.

To make the story nicer, we also define left monad actions (modules) and
characterise Eilenberg-Moore categories as universal left monad actions
along the way.