modify prose: Cat.Functor.Adjoint.Reflective
This file starts out with a diagram L ⊣ ι, but then begins to use the
terms F ⊣ G everywhere. I thought this was a bit confusing, and tried to
make the whole file L ⊣ ι. When used in theorem names I explicitly
changed to "left" and "right", and at the end of the file (whose
assumptions are a general pair of adjoint functors) I changed to the
more clear L ⊣ R.

Other changes are `:'<,'>gq` and minor prose adjustments/fixes.