fix biadjunction diagrams (#532)
Fixes and simplifies the diagrams for adjunctions in a bicategory by
leaving out the coherences. I've also removed the commutative altar (😆)
because it's ambiguous what its commutativity should mean (yes I realise
some of my diagrams are much worse in this respect) and I don't think
it's very important to insert all the coherences anyway.