Homotopy.Join as Logical Disjunction (#546)
Suppose `X` is a type, and we want to produce an element of `X` under
the assumption "`P` or `Q`". Intuitively, we should have to provide an
element under the assumption `P`, provide an element under the
assumption `Q`, and show that the elements agree when both assumptions
hold.

This sounds exactly like the elimination principle for `_∗_` from
`Homotopy.Join`! And indeed, `P ∗ Q` is a proposition that correctly
represents the logical disjunction of `P` and `Q`. I provide the
groundwork required to use it this way.

This could be useful, for example, as a potential resolution to issue
#474. But beyond that, I'd argue that `P ∗ Q` is a good replacement for
`∥ P ⊎ Q ∥` pretty much anywhere. It's equivalent, but it lets you
eliminate into arbitrary types rather than just propositions.

---

In the process of implementing this, I also prove that a join with a
contractible type is always contractible. This fact should be useful on
its own.