chore: make hom[_] a field of Displayed (#554)
Basically everywhere this is a win: we can get the nice computation for
`Gl` directly in terms of `Change-of-base` and `Slice`, and in like…
~98% of our displayed categories, the morphism is only used in a
propositional field. The only exception is `Family` where the base
morphism appears under an arbitrary family of objects so we can't do
much better than `subst` there.

I added two `INLINE` (= becomes copatterns) helpers for defining a
displayed category where `hom[_]` is just `subst`, and one for when the
type of maps is a prop (so you can omit the field). Both of these take
the `Hom-set` field by instance search now.